add set temp by field to cut plot #REF457
